A phone number checking program.
This is a simple phone number checking program, It counts either the phone numbers string length is 11 as per the format of a usual pakistani phone number or not And validate the phone number according it. Then it reads the first four digits of a phone number and tell's who the service provider is.
Note: This program is created only for pakistan based service provider, You can further modify it according to your needs or country standards.